requirements.txt
Flask==3.1.0
langchain==0.3.15
langchain_google_genai==2.0.9
flask-cors==5.0.0
pylint==3.3.3
python-dotenv==1.0.1
